Промежуточное: сделана связь @ManyToOne (Book oo---> Author) на манер как это делают в JPA (ORM, Hibernate) и JPQL

+++ dz08/src/main/java/ru/otus/spring08/dao/IAuthorDao.java
@@ -9,7 +9,7 @@ import java.util.List;
 import ru.otus.spring08.domain.Author;
 
 public interface IAuthorDao {
-  int getAuthorCount();
+  Long getAuthorCount();
   void insertAuthor(Author a);
   Author getById(int id);
   List<Author> getAllAuthors();
